Secure the application

+++ app/api/issues/[id]/route.ts
@@ -1,10 +1,17 @@
+import authOptions from "@/app/auth/authOptions";
 import { issueSchema } from "@/app/validationSchema";
 import { prisma } from "@/prisma/client";
+import { getServerSession } from "next-auth";
 import { NextRequest, NextResponse } from "next/server";
 
 export async function PATCH(
     request: NextRequest, 
     { params }: { params: { id: string }}) {
+        
+    const session = await getServerSession(authOptions);
+    
+    if (!session)
+        return NextResponse.json({}, { status: 401 });
 
     const body = await request.json();
     const validation = issueSchema.safeParse(body);
@@ -34,16 +41,21 @@ export async function DELETE(
     request: NextRequest, 
     { params }: { params: { id: string }}) {
 
-        const issue = await prisma.issue.findUnique({
-            where: { id: parseInt(params.id)}
-        });
+    const session = await getServerSession(authOptions);
 
-        if (!issue)
-            return NextResponse.json({ error: "Invalid issue"}, { status: 404 });
+    if (!session)
+        return NextResponse.json({}, { status: 401 });
 
-        await prisma.issue.delete({
-            where: { id: issue.id }
-        })
+    const issue = await prisma.issue.findUnique({
+        where: { id: parseInt(params.id)}
+    });
 
-        return NextResponse.json({});
-    }
+    if (!issue)
+        return NextResponse.json({ error: "Invalid issue"}, { status: 404 });
+
+    await prisma.issue.delete({
+        where: { id: issue.id }
+    })
+
+    return NextResponse.json({});
+}
